fixed metaFixture errors related to writeOracleEntry (Compound Rate Oracle)

+++ contracts/interfaces/compound/ICToken.sol
@@ -22,7 +22,7 @@ interface ICToken {
     // function borrowBalanceCurrent(address account) external returns (uint);
     // function borrowBalanceStored(address account) external view returns (uint);
     // function exchangeRateCurrent() external returns (uint);
-    function exchangeRateStored() external view returns (uint);
+    function exchangeRateStored() external view returns (uint256);
     // function getCash() external view returns (uint);
     // function accrueInterest() external returns (uint);
     // function seize(address liquidator, address borrower, uint seizeTokens) external returns (uint);
